// src/api/types/CaptchaEventMessage.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

import * as Airtop from "../index";

export interface CaptchaEventMessage {
    /** Time taken to solve the CAPTCHA (in milliseconds). */
    duration: number;
    /** Event name */
    event: string;
    /** Date and Time of the event */
    eventTime: Date;
    /** CAPTCHA job ID */
    id: string;
    /** Retry count */
    retryCount: number;
    /** CAPTCHA solving status */
    status: Airtop.CaptchaEventMessageStatus;
    /** CAPTCHA type */
    type: string;
    /** URL of the page containing the CAPTCHA challenge. */
    url: string;
    /** Window ID */
    windowId: string;
}